Break up the model and components into smaller pieces
