+++ src/component/tocId.tsx
@@ -1,0 +1,4 @@
+
+export const tocId: string = 'table-of-contents';
+
+export default tocId;
