+++ src/model/CentrinelMessage/AnalysisMessageBase.tsx
@@ -1,0 +1,7 @@
+
+import { ErrorLevel } from './ErrorLevel';
+
+export interface AnalysisMessageBase {
+  readonly errorLevel: ErrorLevel;
+  readonly position: string;
+}
