+++ src/model/CentrinelMessage/ErrorLevel.tsx
@@ -1,0 +1,1 @@
+export type ErrorLevel = 'WARNING' | 'ERROR' | 'FATAL ERROR';
